Rename alert app to incident
