--- src/argus/alert/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class AlertConfig(AppConfig):
-    name = "argus.alert"
-    label = "argus_alert"
