+++ src/argus/incident/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "argus.incident.apps.IncidentConfig"
