+++ src/argus/incident/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class IncidentConfig(AppConfig):
+    name = "argus.incident"
+    label = "argus_incident"
