+++ src/argus/incident/fixtures/generate_fixtures.py
@@ -13,14 +13,14 @@ from django.utils import timezone
 
 FIXTURES_DIR = Path(__file__).resolve().parent
 ROOT_DIR = FIXTURES_DIR.parent.parent.parent.parent
-ALERT_FIXTURES_FILE = FIXTURES_DIR / "alert" / "mock_data.json"
+ALERT_FIXTURES_FILE = FIXTURES_DIR / "incident" / "mock_data.json"
 
 # Must be called before any models are imported, to be able to use them
 sys.path.append(ROOT_DIR)
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "argus.site.settings.dev")
 django.setup()
 
-from argus.alert.models import *
+from argus.incident.models import *
 
 
 # --- Generation configuration ---
