+++ src/argus/incident/mappings.py
