+++ src/argus/incident/models.py
