+++ src/argus/incident/utils.py
