+++ src/argus/incident/views.py
