+++ src/argus/notificationprofile/notification_media.py
@@ -7,8 +7,8 @@ from django.conf import settings
 from django.template.loader import render_to_string
 from rest_framework.renderers import JSONRenderer
 
-from argus.alert.models import Alert
-from argus.alert.serializers import AlertSerializer
+from argus.incident.models import Alert
+from argus.incident.serializers import AlertSerializer
 from argus.auth.models import User
 from .models import NotificationProfile
